Extract and create ProjectsComponent
