+++ src/app/modules/data-set/projects/projects.component.html
@@ -1,0 +1,100 @@
+<div class="table-container mat-elevation-z8 container">
+    <p class="table-title">Projects</p>
+    <div class="table-header">
+        <mat-form-field appearance="fill" class="search">
+            <mat-label>Search by name</mat-label>
+            <input matInput (keyup)="searchProjects($event)" placeholder="Project name">
+        </mat-form-field>
+        <button class="add-button" mat-mini-fab matTooltip="New project" matTooltipPosition="right" (click)="addProject()" color="primary">
+            <mat-icon>add</mat-icon>
+        </button>
+    </div>
+    <div id="projects-table-div">
+        <table mat-table [dataSource]="dataSource" class="mat-elevation-z8">
+            <ng-container matColumnDef="name">
+                <th mat-header-cell *matHeaderCellDef width="30%"> Name </th>
+                <td mat-cell *matCellDef="let project" width="30%"> 
+                    <span class="bold-name" (click)="project.state != projectState.Failed && chooseProject(project.id)">{{project.name}}</span>
+                </td>
+            </ng-container>
+            <ng-container matColumnDef="url">
+                <th mat-header-cell *matHeaderCellDef width="10%"> URL </th>
+                <td mat-cell *matCellDef="let project" width="10%">
+                    <a href="{{project.url}}" target="_blank">
+                        <mat-icon>open_in_new</mat-icon>
+                    </a>
+                </td>
+            </ng-container>
+            <ng-container matColumnDef="numOfInstances">
+                <th mat-header-cell *matHeaderCellDef width="5%" class="cell-center">Instances</th>
+                <td mat-cell *matCellDef="let project" width="5%" class="cell-center">{{project.instancesCount}}</td>
+            </ng-container>
+            <ng-container matColumnDef="fullyAnnotated">
+                <th mat-header-cell *matHeaderCellDef class="cell-center" width="10%" >Fully annotated by me</th>
+                <td mat-cell *matCellDef="let project" class="cell-center" width="10%" >
+                    <ng-container *ngIf="project.state == projectState.Built && project.fullyAnnotated">
+                        <mat-icon class="icon-built-project">check_circle_outline</mat-icon>
+                    </ng-container>
+                    <ng-container *ngIf="project.state == projectState.Built && !project.fullyAnnotated">
+                        <mat-icon>highlight_off</mat-icon>
+                    </ng-container>
+                    <ng-container *ngIf="project.state == projectState.Failed">
+                        <mat-icon>highlight_off</mat-icon>
+                    </ng-container>
+                </td>
+            </ng-container>
+            <ng-container matColumnDef="consistency">
+                <th mat-header-cell *matHeaderCellDef width="5%" class="cell-center">Annotation consistency</th>
+                <td mat-cell *matCellDef="let project" width="5%" class="cell-center">
+                    <ng-container *ngIf="project.instancesCount > 0">
+                        <button mat-stroked-button color="primary" (click)="checkConsistency(project.id)">Check</button>
+                    </ng-container>
+                </td>
+            </ng-container>
+            <ng-container matColumnDef="projectUpdate">
+                <th mat-header-cell *matHeaderCellDef width="5%" ></th>
+                <td mat-cell *matCellDef="let project" width="5%">
+                    <button mat-icon-button (click)="updateProject(project)" color="primary">
+                        <mat-icon matTooltip="edit">edit</mat-icon>
+                    </button>
+                </td>
+            </ng-container>
+            <ng-container matColumnDef="projectDelete">
+                <th mat-header-cell *matHeaderCellDef width="5%"></th>
+                <td mat-cell *matCellDef="let project" width="5%">
+                    <button mat-icon-button (click)="deleteProject(project.id)" color="primary">
+                        <mat-icon matTooltip="delete">delete</mat-icon>
+                    </button>
+                </td>
+            </ng-container>
+            <ng-container matColumnDef="status">
+                <th mat-header-cell *matHeaderCellDef width="5%" class="cell-center">Build status</th>
+                <td mat-cell *matCellDef="let project" width="5%" class="cell-center">
+                    <ng-container *ngIf="project.state == projectState.Processing">
+                        <mat-spinner matTooltip="processing" diameter="20">
+                        </mat-spinner>
+                    </ng-container>
+                    <ng-container *ngIf="project.state == projectState.Built">
+                        <mat-icon matTooltip="built" class="icon-built-project">check_circle_outline</mat-icon>
+                    </ng-container>
+                    <ng-container *ngIf="project.state == projectState.Failed">
+                        <mat-icon matTooltip="failed" class="icon-failed-project">highlight_off</mat-icon>
+                    </ng-container>
+                </td>
+            </ng-container>
+            <tr mat-header-row *matHeaderRowDef="displayedColumns; sticky: true"></tr>
+            <tr mat-row *matRowDef="let row; columns: displayedColumns;" [class.row-chosen]="row.id == chosenProject.id"></tr>
+        </table>
+    </div>
+
+    <form class="filterForm">
+        <mat-form-field appearance="fill" *ngIf="chosenProject?.id">
+            <mat-label>Choose instances filter</mat-label>
+            <mat-select required name="instancesFilter" (selectionChange)="chooseProject(chosenProject.id)" [formControl]="filterFormControl">
+            <mat-option *ngFor="let filter of instancesFilters" [value]="filter" >
+                {{filter}}
+            </mat-option>
+            </mat-select>
+        </mat-form-field>
+    </form>
+</div>
